# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가 시스템 공통기능 API
  version: 1.0.0
  description: 1차 시스템 관리 영역의 HTTP API 계약이다.
servers:
- url: /
  description: 브라우저 상대경로 API 기준
paths:
...
  /api/menus:
    get:
      operationId: listMenus
      summary: 메뉴 실행정보 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-019
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      x-roles:
      - R09
  /api/menus/{menuId}:
    put:
      operationId: updateMenu
      summary: 메뉴 실행정보 변경
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-020
      - REQ-021
      - REQ-028
      - REQ-029
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: menuId
        in: path
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuUpdateRequest'
      x-roles:
      - R09
      x-business-rules:
      - 메뉴명·화면ID·URL·아이콘·업무구분·설명만 실행정보로 저장한다
      x-side-effects:
      - menu 실행정보와 화면 연결을 저장한다
      - change_history에 실행정보 변경을 기록한다
      x-required-tests:
      - 'happy: 유효한 R09 요청 -> menu의 screen_id 변경을 후속 GET으로 확인'
      - 'auth: R09 이외의 인증 사용자 요청 -> 403과 menu 미변경'
      - 'validation: 필수 screen_id 누락 -> ApiError fieldErrors에 screen_id 오류'
      - 'business: 메뉴-실행 화면 연결 값 검증 위반 요청 -> 400과 대상 상태 미변경'
      - 'side-effect: 성공 요청 -> change_history에 메뉴 실행정보 변경 이력 존재'
...
  /api/code-groups/{groupId}/detail-codes:
    get:
      operationId: listDetailCodes
      summary: 상세코드 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-024
      - REQ-025
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: groupId
        in: path
        required: true
        schema: *id002
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      x-roles:
      - R09
  /api/detail-codes:
    post:
      operationId: createDetailCode
      summary: 상세코드 등록
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-026
      - REQ-027
      - REQ-028
      - REQ-029
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailCodeRequest'
      x-roles:
      - R09
      x-business-rules:
      - group_id 안에서 code_value는 중복될 수 없다
      - 상위코드는 같은 그룹의 상세코드여야 한다
      x-side-effects:
      - detail_code를 ACTIVE로 생성한다
      - change_history에 상세코드 등록을 기록한다
      x-required-tests:
      - 'happy: 유효한 R09 요청 -> detail_code의 code_value 변경을 후속 GET으로 확인'
      - 'auth: R09 이외의 인증 사용자 요청 -> 403과 detail_code 미변경'
      - 'validation: 필수 code_value 누락 -> ApiError fieldErrors에 code_value 오류'
      - 'business: 같은 그룹 code_value 중복 금지 위반 요청 -> 400과 대상 상태 미변경'
      - 'side-effect: 성공 요청 -> change_history에 상세코드 변경 이력 존재'
  /api/detail-codes/{detailCodeId}:
    put:
      operationId: updateDetailCode
      summary: 상세코드 변경
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-026
      - REQ-027
      - REQ-028
      - REQ-029
      responses:
        '200': *id001
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: detailCodeId
        in: path
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailCodeRequest'
      x-roles:
      - R09
      x-business-rules:
      - 상위코드는 자신을 가리킬 수 없다
      - 상위코드는 같은 그룹의 상세코드여야 한다
      x-side-effects:
      - detail_code의 코드명·정렬순서·추가속성을 수정한다
      - change_history에 상세코드 변경을 기록한다
      x-required-tests:
      - 'happy: 유효한 R09 요청 -> detail_code의 parent_detail_code_id 변경을 후속 GET으로 확인'
      - 'auth: R09 이외의 인증 사용자 요청 -> 403과 detail_code 미변경'
      - 'validation: 필수 parent_detail_code_id 누락 -> ApiError fieldErrors에 parent_detail_code_id 오류'
      - 'business: 자기 자신 상위코드 금지 위반 요청 -> 400과 대상 상태 미변경'
      - 'side-effect: 성공 요청 -> change_history에 상세코드 변경 이력 존재'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: HttpOnly SameSite=Lax 세션 cookie
  schemas:
...
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
        meta:
          type: object
          additionalProperties: true
...
    MenuUpdateRequest:
      type: object
      required:
      - menuName
      - screenId
      - url
      - reason
      properties:
        menuName:
          type: string
        screenId:
          type: string
        url:
          type: string
        icon:
          type: string
        businessCategory:
          type: string
        description:
          type: string
        reason:
          type: string
...
    DetailCodeRequest:
      type: object
      required:
      - groupId
      - codeValue
      - codeName
      - displayOrder
      - reason
      properties:
        groupId:
          type: string
        codeValue:
          type: string
        codeName:
          type: string
        parentDetailCodeId:
          type: string
          format: uuid
        displayOrder:
          type: integer
          minimum: 0
        additionalAttributes:
          type: object
          additionalProperties: true
        reason:
          type: string
...